# Marshal directly to bytebuffer where possible.