Uses a new API introduced in 1.39 to let us write straight to a netty buffer instead of going through a wasteful buffered outputstream

After
```
Benchmark                                                                         (numSpans)  Mode  Cnt      Score      Error   Units
MarshalerInputStreamBenchmarks.marshalToNettyBuffer                                       16  avgt   10     15.713 ±    0.133   us/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate                        16  avgt   10    916.361 ±    7.812  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate.norm                   16  avgt   10  22678.477 ±    0.554    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space               16  avgt   10    923.821 ±   95.072  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space.norm          16  avgt   10  22861.759 ± 2312.486    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen                  16  avgt   10      0.015 ±    0.006  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen.norm             16  avgt   10      0.372 ±    0.148    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.count                             16  avgt   10     76.000             counts
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.time                              16  avgt   10     33.000                 ms
```

Before
```

Benchmark                                                                         (numSpans)  Mode  Cnt      Score      Error   Units
MarshalerInputStreamBenchmarks.marshalToNettyBuffer                                       16  avgt   10     18.072 ±    0.390   us/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate                        16  avgt   10    936.069 ±   20.049  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.alloc.rate.norm                   16  avgt   10  26638.468 ±    0.297    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space               16  avgt   10    935.413 ±   88.690  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Eden_Space.norm          16  avgt   10  26621.998 ± 2515.013    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen                  16  avgt   10      0.015 ±    0.007  MB/sec
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.churn.G1_Old_Gen.norm             16  avgt   10      0.432 ±    0.212    B/op
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.count                             16  avgt   10     77.000             counts
MarshalerInputStreamBenchmarks.marshalToNettyBuffer:·gc.time                              16  avgt   10     32.000                 ms
```