@anuraaga 2021-08-19
Nevermind I interpreted the release notes wrong, this interface is only used on parsing not writing. Maybe in the future then